[Chore] Make folder structure
